atomic-examples/nested-multimodal/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nested_multimodal"]

[project]
name = "nested-multimodal"
version = "1.0.0"
description = "Nested multimodal example demonstrating images/PDFs inside nested Pydantic schemas"
readme = "README.md"
authors = [
    { name = "Kenny Vaneetvelde", email = "kenny.vaneetvelde@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "atomic-agents",
    "instructor==1.13.0",
    "openai>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[tool.uv.sources]
atomic-agents = { workspace = true }
